Change operator logic
Only a maximum of two operands can be operated at any given step.

As per the requirements, the calculation is now continuous, which means
that if an operator like "+" is entered for a second time instead of
the "=" sign, the result is displayed and the operator saved, so that
the next number entered will be operated with the result.